+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Version 1.0.dev
 
 - Add .gitkeep to .gitignore.
+- Add scripts directory.
+- Change to `/vagrant` on login.
+- Add puppet-lint.
+- Add git.
+- Add mercurial.
+- Add vim.
 
 ## Version 1.0